add more logging at info level for development
added logging to graphite and carbon plugins to log some calls and
log every call to cassandra.

The carbon plugin just gets the pyton logger with the same name
graphite does, so it can log to its file.

Logs are written at info level, so are in

  /opt/graphite/storage/log/webapp/info.log

Navigate the "Graphite" node in the tree and watch those DB calls fly.