Squash commits for review
Fix typo, add class docstring.

Fix Exception definition.

Write metadata to CF in key-value format.

Still seeing an error when trying to retrieve DataNodes.

Return formatted columns.

Write out datapoints to cf in key-value format.

Convert column values only once per datanode.

DataNode.fromCols() is called when DataNode.__init__() is invoked,
and the same data node can be `init`ed multiple times.

Perhaps in the future we will want to set a 'converted' flag to
identify whether data has been transformed or not, but for now,
pass back the data if it is in the correct format.

Also, a couple of typos were fixed as well.