Convert column values only once per datanode.
DataNode.fromCols() is called when DataNode.__init__() is invoked,
and the same data node can be `init`ed multiple times.

Perhaps in the future we will want to set a 'converted' flag to
identify whether data has been transformed or not, but for now,
pass back the data if it is in the correct format.

Also, a couple of typos were fixed as well.