Replicate the node tree per DC.
Added carbon writer config LOCAL_DC_NAME to db.conf. This is passed
to the DataTree so it knows which DC it is running in. When a node is
created the node tree is replicated into a column family called
"dc_$DCNAME_nodes". e.g. "dc_us-west-1_nodes"

Updated DataNode.selfAndChildPaths() to accept a DC name so we can
segment maintenance (e.g. rollup) operations per DC.