Handle missing data points
Ceres zero pads it’s slice files on disk when there is a missing data
point. The start time and time step are used to calculate the timestamp
for each fixed width data point on disk. When writing a data point if
the gap is too big from the last one (default is 10 points) a new slice
file is created.

With cassandra ave a single Data Slice / row per Data Node. We store
all points in this row and do not write missing data points. This was
causing problems during rollups because a set of points read from disk
between A and B would have missing data.

Added TimeSeriesData.fromDB() which None pads the data to match Ceres.

Tidied up DataNode.read() and made a change. Changed the first break to
continue in the loop, see comments. If the first DataSlice had no data
for the from / until range it was not looking in other DataRanges.